Move HTTPURLResponse(next:) to TestsFoundation
